+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
+## 1.0.0
+- Cleanup Syntax of Loading / Reading Config
+- Remove all get_(bool, int, float)
+- Add get_str for getting raw config value
+- add get<T> for gatting a config value as any type (that impls FromStr)
+- Update Exmaples
+
 ## 0.1.6
 - Add docs for internal functions
 - Update some formatting in the readme